ensure we only register each R version once
